always open links in new tabs, rename popover
