+++ client/web/src/components/Dialog.scss
@@ -6,8 +6,7 @@
     width: 32rem;
     max-width: 100vw;
 
-    &__center {
-        // Center modal
+    &--centered {
         transform: translate(-50%, -50%);
         position: absolute;
         top: 50%;
