+++ client/web/src/repo/actions/GoToCodeHostAction.tsx
@@ -15,7 +15,7 @@ import { ExternalLinkFields } from '../../graphql-operations'
 import { useObservable } from '../../../../shared/src/util/useObservable'
 import GitlabIcon from 'mdi-react/GitlabIcon'
 import { eventLogger } from '../../tracking/eventLogger'
-import { InstallExtensionPopover } from './InstallExtensionPopover'
+import { InstallBrowserExtensionPopover } from './InstallBrowserExtensionPopover'
 
 interface GoToCodeHostPopoverProps {
     canShowPopover: boolean
@@ -107,9 +107,6 @@ export const GoToCodeHostAction: React.FunctionComponent<Props> = props => {
         }
     }, [hijackLink, showPopover])
 
-    // Keep track of latest click type for install extension popover
-    const [wasAuxClick, setWasAuxClick] = useState(false)
-
     const onClick = useCallback(
         (event: React.MouseEvent<HTMLAnchorElement, MouseEvent>) => {
             if (showPopover) {
@@ -120,24 +117,7 @@ export const GoToCodeHostAction: React.FunctionComponent<Props> = props => {
 
             if (hijackLink) {
                 event.preventDefault()
-                setWasAuxClick(false)
-                setShowPopover(true)
-            }
-        },
-        [hijackLink, showPopover]
-    )
 
-    const onAuxClick = useCallback(
-        (event: React.MouseEvent<HTMLAnchorElement, MouseEvent>) => {
-            if (showPopover) {
-                event.preventDefault()
-                setShowPopover(false)
-                return
-            }
-
-            if (hijackLink) {
-                event.preventDefault()
-                setWasAuxClick(true)
                 setShowPopover(true)
             }
         },
@@ -211,12 +191,11 @@ export const GoToCodeHostAction: React.FunctionComponent<Props> = props => {
                 data-tooltip={`View on ${displayName}`}
                 id={TARGET_ID}
                 onClick={onClick}
-                onAuxClick={onAuxClick}
             >
                 <Icon className="icon-inline" />
             </a>
 
-            <InstallExtensionPopover
+            <InstallBrowserExtensionPopover
                 url={url}
                 toggle={toggle}
                 isOpen={showPopover}
@@ -225,7 +204,6 @@ export const GoToCodeHostAction: React.FunctionComponent<Props> = props => {
                 onRejection={onRejection}
                 onClickInstall={onClickInstall}
                 targetID={TARGET_ID}
-                wasAuxClick={wasAuxClick}
             />
         </>
     )
