+++ client/web/src/repo/actions/InstallBrowserExtensionPopover.scss
