+++ client/web/src/repo/actions/InstallBrowserExtensionPopover.story.tsx
